requirements.txt
flask==2.3.2
python-dotenv      
openai